add all model and seed

+++ database/migrations/2021_01_11_064813_create_vanchuyen.php
@@ -18,7 +18,7 @@ class CreateVanchuyen extends Migration
             $table->engine = 'InnoDB';
             $table->unsignedTinyInteger('vc_ma')->autoIncrement()->comment('Mã dịch vụ giao hàng');
             $table->string('vc_ten', 200)->comment('Tên dịch vụ # tên dịch vụ giao hàng');
-            $table->tinyInteger('vc_chiPhi')->default('0')->comment('Phí giao hàng # Phí giao hàng');
+            $table->float('vc_chiPhi',2)->default('0')->comment('Phí giao hàng # Phí giao hàng');
             $table->text('vc_dienGiai')->comment('Thông tin # thông tin dịch vụ giao hàng');
             $table->timestamp('vc_taoMoi')->default(DB::raw('CURRENT_TIMESTAMP'))->comment('Thời điểm tạo # Thời điểm đầu tiên tạo dịch vụ giao hàng');
             $table->timestamp('vc_capNhat')->default(DB::raw('CURRENT_TIMESTAMP'))->comment('Thời điểm cập nhật # Thời điểm cập nhật dịch vụ giao hàng gần nhất');
